backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
psycopg[binary]==3.2.12
python-dotenv==1.0.1
requests==2.32.3
bcrypt==4.1.2
